+++ myconext-server/src/main/resources/application.yml
@@ -90,6 +90,8 @@ mobile_app_rp_entity_id: mobile_app_rp_entity_id
 host_headers:
   service_desk: servicedesk.test2.eduid.nl
   mijn_ediuid: mijn.test2.eduid.nl
+  #Only for testing purposes
+  active: mijn.test2.eduid.nl
 
 feature:
   webauthn: True
