+++ myconext-server/src/test/java/myconext/AbstractIntegrationTest.java
@@ -90,7 +90,8 @@ import static org.junit.Assert.assertTrue;
                 "sso_mfa_duration_seconds=-1000",
                 "feature.requires_signed_authn_request=false",
                 "feature.deny_disposable_email_providers=false",
-                "verify.base_uri=http://localhost:8098"
+                "verify.base_uri=http://localhost:8098",
+                "host_headers.active=mijn.test2.eduid.nl"
         })
 @ActiveProfiles({"test"})
 @SuppressWarnings("unchecked")
