+++ myconext-server/src/test/java/myconext/api/ServiceDeskControllerTest.java
@@ -21,7 +21,8 @@ import static org.junit.jupiter.api.Assertions.*;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
         properties = {
                 "service_desk_role_auto_provisioning=True",
-                "feature.service_desk_active=False"
+                "feature.service_desk_active=True",
+                "host_headers.active=servicedesk.test2.eduid.nl"
         })
 class ServiceDeskControllerTest extends AbstractIntegrationTest {
 
@@ -131,16 +132,4 @@ class ServiceDeskControllerTest extends AbstractIntegrationTest {
                 .statusCode(403);
     }
 
-    @Test
-    void createUserControlCode() {
-        clearExternalAccounts("jdoe@example.com");
-        ControlCode controlCode = new ControlCode("Lee", "Harpers", "01 Mar 1977");
-        given()
-                .body(controlCode)
-                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/myconext/api/sp/control-code")
-                .then()
-                .statusCode(HttpStatus.FORBIDDEN.value());
-    }
 }
