+++ myconext-server/src/test/java/myconext/api/ServiceDeskControllerUnauthorizedTest.java
@@ -1,8 +1,11 @@
 package myconext.api;
 
 import myconext.AbstractIntegrationTest;
+import myconext.model.ControlCode;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 
 import static io.restassured.RestAssured.given;
 
@@ -20,4 +23,17 @@ class ServiceDeskControllerUnauthorizedTest extends AbstractIntegrationTest {
                 .then()
                 .statusCode(403);
     }
+
+    @org.junit.Test
+    public void createUserControlCodeForbidden() {
+        ControlCode controlCode = new ControlCode("Lee", "Harpers", "01 Mar 1977");
+        given()
+                .body(controlCode)
+                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/myconext/api/sp/control-code")
+                .then()
+                .statusCode(403);
+    }
+
 }
